chore: apply forge fmt

+++ test/FANtiumUserManagerV2.t.sol
@@ -215,10 +215,10 @@ contract FANtiumUserManagerV2Test is Test {
     )
         public
     {
-        uint max = type(uint256).max;
+        uint256 max = type(uint256).max;
         vm.assume(account != address(0));
         vm.assume(delta > 0);
-        initialAmount = bound(initialAmount,  max - delta, max);
+        initialAmount = bound(initialAmount, max - delta, max);
 
         vm.startPrank(allowlistManager);
         userManager.setAllowList(account, collectionId, initialAmount);
